services/backend/identity-service/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
